+++ games_services/lib/src/games_services.dart
@@ -92,13 +92,20 @@ class GamesServices {
 
   /// Open the device's default leaderboards screen. If a leaderboard ID is provided,
   /// it will display the specific leaderboard, otherwise it will show the list of all leaderboards.
+  ///
+  /// The `timeScope` parameter allows you to specify the time range for the leaderboard scores, 
+  /// only supported on Android. It has no effect on iOS.
   static Future<String?> showLeaderboards({
     String iOSLeaderboardID = "",
     String androidLeaderboardID = "",
+    TimeScope timeScope = TimeScope.allTime,
+    PlayerScope playerScope = PlayerScope.global,
   }) async {
     return await Leaderboards.showLeaderboards(
-      iOSLeaderboardID: iOSLeaderboardID,
-      androidLeaderboardID: androidLeaderboardID,
+        iOSLeaderboardID: iOSLeaderboardID,
+        androidLeaderboardID: androidLeaderboardID,
+      timeScope: timeScope,
+      playerScope: playerScope,
     );
   }